Move forest-related files into own directory
